requirements.txt
google-generativeai
chromadb
python-dotenv
pandas